Replace splitting by `" "` with regex splitting
The original splitting method resulted in an `IndexError` when using the
stutter flag on a string that contains leading whitespaces. It also
doesn't preserve the spaces in the original string after uwuification.

This commit replaces that splitting method with a regex which prevents
the `IndexError` and also preserves all the whitespaces in the original string.