Don't count spaces or "" when stuttering
Previous commit prevented empty strings or whitespaces from being
stuttered but they were still counted as a normal word. This means that
if a whitespace or empty string was one of the `count_every_nth_word`
strings, then a stutter would be skipped.
This commit introduces a running variable that keeps track of actual
words and excludes whitespaces and empty strings from being counted.